Java Project  Continent and Cities
Switch: Continents and Cities
1.
Create an int variable called continent and set it equal to 4.
2.
Create a switch statement that will switch based on the value of continent.
3.
Inside of the switch statement, add a case that will run when the value of continent is 1.
4.
When the value of continent is 1, print out North America: Mexico City, Mexico.
5.
Make sure the next line exits out of the case.
6.
Add another case that will run when the value of continent is 2. When this value is met, print out South America: Sao Paulo, Brazil.
7.
Make sure the next line exits out of the case.
8.
Add another case that will run when the value of continent is 3. When this value is met, print out Europe: Moscow, Russia. Make sure the next line exits out of the case.
9.
Add another case that will run when the value of continent is 4. When this value is met, print out Africa: Lagos, Nigeria. Make sure the next line exits out of the case.
10.
Add another case that will run when the value of continent is 5. When this value is met, print out Asia: Shanghai, China. Make sure the next line exits out of the case.
11.
Add another case that will run when the value of continent is 6. When this value is met, print out Australia: Sydney, Australia. Make sure the next line exits out of the case.
12.
Add another case that will run when the value of continent is 7. When this value is met, print out Antarctica: McMurdo Station, US. Make sure the next line exits out of the case.
13.
Finally, add the default case. The default case should print out Undefined continent! Make sure the next line exits out of the case.
14.
If the program is written correctly, your output should be Africa: Lagos, Nigeria. Great work!